% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.aggregate_search_terms}
\alias{.aggregate_search_terms}
\title{Create list of search terms}
\usage{
.aggregate_search_terms(search_terms)
}
\value{
a list of all of the given search terms, with 'q' as the name for each of them
}
\description{
Aggregate vector of search terms to build the API call
}
