% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{.build_params}
\alias{.build_params}
\title{Ensures all parameters are in the proper format}
\usage{
.build_params(...)
}
\arguments{
\item{\dots}{Whatever params are given}
}
\value{
all of the necessary parameters
}
\description{
Makes sure all necessary params are there and all vector params are comma-separated
}
