% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{.build_query}
\alias{.build_query}
\title{Helper function that puts together all the necessary parameters}
\usage{
.build_query(type, search_terms, ...)
}
\arguments{
\item{type}{the type of query}

\item{search_terms}{character vector of key words to search for}

\item{\dots}{all of the other pieces of the query}
}
\description{
Helper function that puts together all the necessary parameters
}
