% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{.build_url}
\alias{.build_url}
\title{Uses parameters to build the URL the request will hit}
\usage{
.build_url(path, params)
}
\arguments{
\item{path}{the path of the query}

\item{params}{named list of all of the other parameters (key words, size, etc...)}
}
\description{
Uses parameters to build the URL the request will hit
}
