% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{.parse_search_terms}
\alias{.parse_search_terms}
\title{Makes sure search terms are in the proper format}
\usage{
.parse_search_terms(search_terms)
}
\arguments{
\item{search_terms}{character vector of the search terms}
}
\value{
correctly formatted search terms
}
\description{
Search terms that have spaces or commas are quotes
}
