% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPushshiftData.R
\name{getPushshiftData}
\alias{getPushshiftData}
\title{Gets the pushshift data}
\usage{
getPushshiftData(postType, ...)
}
\arguments{
\item{postType}{One of `submission` or `comment`}

\item{title}{A string to search for in post titles}

\item{size}{Number of results to return, maximum is 1000}

\item{q}{A query to search for}

\item{after}{Only search for posts made after this data, specified as a UNIX epoch time}

\item{before}{As `after`, but before}

\item{subreddit}{Only return posts made in this subreddit}

\item{nest_level}{How deep to search? `nest_level = 1` returns only top-level comments}
}
\value{
A tibble of reddit submissions
}
\description{
Gets the pushshift data
}
