% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPushshiftData.R
\name{getPushshiftDataRecursive}
\alias{getPushshiftDataRecursive}
\title{Repeats getPushshiftData until desired period covered}
\usage{
getPushshiftDataRecursive(
  postType = "submission",
  title = NULL,
  size = NULL,
  q = NULL,
  after = NULL,
  before = NULL,
  subreddit = NULL,
  nest_level = NULL,
  delay = 0
)
}
\arguments{
\item{postType}{One of `submission` or `comment`}

\item{title}{A string to search for in post titles}

\item{size}{Number of results to return, maximum is 1000}

\item{q}{A query to search for}

\item{after}{Only search for posts made after this data, specified as a UNIX epoch time}

\item{before}{As `after`, but before}

\item{subreddit}{Only return posts made in this subreddit}

\item{nest_level}{How deep to search? `nest_level = 1` returns only top-level comments}

\item{delay}{Number of seconds to wait between queries to avoid stressing out the Pushshift server (limit is somewhere around 200 queries per minute)}
}
\value{
A tibble of the requested data
}
\description{
Repeats getPushshiftData until desired period covered
}
