% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{This is my favorite function.}
\usage{
... \%notin\% NA
}
\arguments{
\item{...}{two vectors: one that you're checking the values for and one that you're checking against}
}
\description{
The exact opposite of `%in%``
}
\details{
equivalent to \code{!x \%in\% y}
}
