% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{new_ps_query}
\alias{new_ps_query}
\title{Structured way to create to properly create a ps_query object}
\usage{
new_ps_query(url, type, sort_type, sort_direction, pagination_strategy,
  size)
}
\arguments{
\item{url}{the full URL endpoint the query is going to hit (params and all)}

\item{type}{the type of query}

\item{sort_type}{whether to sort on time, score, etc...}

\item{sort_direction}{ascending or descending}

\item{pagination_strategy}{different ways to paginate}

\item{size}{the number of rows returned per request}
}
\description{
Structured way to create to properly create a ps_query object
}
