% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{param_to_csv}
\alias{param_to_csv}
\title{Turn character vectors into one comma separated string}
\usage{
param_to_csv(value)
}
\arguments{
\item{value}{either a character vector or list of character vectors}
}
\value{
value modified such that each individual character vector is now one string with comma separated values
}
\description{
Turn character vectors into one comma separated string
}
