% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ps_get_comment_ids}
\alias{ps_get_comment_ids}
\title{Search reddit comments}
\usage{
ps_get_comment_ids(submission_id)
}
\arguments{
\item{submission_id}{the id of the submission (post) in question.}
}
\description{
Gets the comment ids for the comments in specific posts
}
