% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ps_query}
\alias{ps_query}
\title{Creates a Query S3 Object from many distinct pieces.}
\usage{
ps_query(type, ...)
}
\arguments{
\item{type}{the type of query (subreddit, post, comment, etc...)}

\item{\dots}{all of the other arguments}
}
\description{
Pass in the type of query and all the other varied arguments that you need and it will
create a proper ps_query object for you.
}
