% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ps_search}
\alias{ps_search}
\title{Using the API for search}
\usage{
ps_search(type = c("comment", "submission", "subreddit"), ...)
}
\arguments{
\item{type}{Either "comment" or "submission", depending on which objects you want to search.}

\item{\dots}{All of the other search parameters (e.g. size, before, until, etc...)}
}
\value{
a dataframe of the objects that the API has returned
}
\description{
Builds the query from parameters that you pass in, handles pagination, etc...
}
