% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ps_search_comments}
\alias{ps_search_comments}
\title{Search reddit comments}
\usage{
ps_search_comments(search_terms = "", ...)
}
\arguments{
\item{search_terms}{character vector of the key words to search for}

\item{...}{Any extra parameters to give the query, e.g. since, until, etc...}
}
\description{
High level function that lets you easily search comments
}
