% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ps_search_submissions}
\alias{ps_search_submissions}
\title{Search submissions}
\usage{
ps_search_submissions(search_terms = "", ...)
}
\description{
High level function that lets you easily search submissions.
}
