% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{search_reddit}
\alias{search_reddit}
\title{Search reddit}
\usage{
search_reddit(content, ...)
}
\arguments{
\item{content}{Character. Either "comments" or "submissions"}

\item{...}{The parameters to the search. Unnamed parameters are assumed to be search terms.}
}
\value{
A dataframe of the results.
}
\description{
High level function that lets you easily search submissions or comments.
}
