% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.aggregate_search_terms}
\alias{.aggregate_search_terms}
\title{Create list of search terms
Deprecated}
\usage{
.aggregate_search_terms(search_terms)
}
\arguments{
\item{search_terms}{a character vector of search terms that will be matched against the text of the submissions or comments.}
}
\value{
a list of all of the given search terms, with 'q' as the name for each of them
}
\description{
Aggregate vector of search terms to build the API call
}
