% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.build_params}
\alias{.build_params}
\title{Builds the list of query parameters}
\usage{
.build_params(...)
}
\arguments{
\item{...}{the list of parameters. Non-named parameters are assumed to be search terms.}
}
\value{
list of parameters
}
\description{
Puts together the list of query parameters necessary to query the API.
}
\details{
There are two necessary parameters: "size" and "sort". If they are not present, then they
are added. Their default values are 25 and "desc" (i.e. "descending"), respectively.
}
