% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{The opposite of %in%}
\usage{
x %in% y
}
\arguments{
\item{x}{vector, the values to be (not) matched}

\item{y}{vector, the values to be (not) matched against}
}
\arguments{
\item{...}{two vectors: one that you're checking the values for and one that you're checking against}
}
\description{
The opposite of %in%
}
\examples{
4 \%notin\% c(1:3, 5:10)
"hello" \%notin\% c("hello", "it's", "me")
}
