% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ps_reddit_api}
\alias{ps_reddit_api}
\title{Basic function to query the pushshift.io API}
\usage{
ps_reddit_api(url, previously_provided_ids = character(0))
}
\arguments{
\item{url:}{the pushshift.io URL that contains the query}

\item{previously_provided_ids:}{IDs of comments/submissions that have already been returned}
}
\value{
pushshift_api object, includes the response headers, the content as a dataframe, and the min/max utc of the objects in the response.
}
\arguments{
\item{url}{httr URL object, contains all of the info about which endpoint, search terms, etc...}

\item{provided_ids}{character vector of ids to exclude}
}
\description{
Given a URL, returns the response
}
