% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ps_reddit_api}
\alias{ps_reddit_api}
\title{Basic function to query the pushshift.io API}
\usage{
ps_reddit_api(url, provided_ids = character(0))
}
\arguments{
\item{url}{httr URL object, contains all of the info about which endpoint, search terms, etc...}

\item{provided_ids}{character vector of ids to exclude}
}
\description{
Given a URL, returns the response
}
