% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ps_search_submissions}
\alias{ps_search_submissions}
\title{Search submissions}
\usage{
ps_search_submissions(search_terms = "", ...)
}
\arguments{
\item{search_terms}{The terms to search for. Will search the post title and selftext.}

\item{\dots}{Whatever other query parameters that are provided (e.g. size, before, until, etc...)}
}
\value{
a dataframe
}
\description{
High level function that lets you easily search submissions.
}
